+++ packages/sdk/src/plt/Token.ts
@@ -10,7 +10,7 @@ import {
     TransactionHash,
 } from '../pub/types.js';
 import { AccountSigner, signTransaction } from '../signHelpers.js';
-import { Cbor, TokenAmount, TokenId, TokenInfo, TokenModuleReference } from './index.js';
+import { Cbor, TokenAmount, TokenHolder, TokenId, TokenInfo, TokenModuleReference } from './index.js';
 import {
     TokenAddAllowListOperation,
     TokenAddDenyListOperation,
@@ -130,9 +130,9 @@ export class NotAllowedError extends TokenError {
 
     /**
      * Constructs a new NotAllowedError.
-     * @param {AccountAddress.Type} receiver - The account address of the receiver.
+     * @param {TokenHolder.Type} receiver - The account address of the receiver.
      */
-    constructor(public readonly receiver: AccountAddress.Type) {
+    constructor(public readonly receiver: TokenHolder.Type) {
         super(
             `Transfering funds from or to the account specified is currently not allowed (${receiver}) because of the allow/deny list.`
         );
@@ -335,9 +335,10 @@ export async function validateTransfer(
     payloads.forEach((p) => validateAmount(token, p.amount));
 
     const { decimals } = token.info.state;
+    const senderInfo = await token.grpc.getAccountInfo(sender);
 
     // Check the sender balance.
-    const senderBalance = (await balanceOf(token, sender)) ?? TokenAmount.zero(decimals); // We fall back to zero, as the `token` has already been validated at this point.
+    const senderBalance = balanceOf(token, senderInfo) ?? TokenAmount.zero(decimals); // We fall back to zero, as the `token` has already been validated at this point.
     const payloadTotal = payloads.reduce(
         (acc, { amount }) => acc.add(TokenAmount.toDecimal(amount)),
         TokenAmount.toDecimal(TokenAmount.zero(decimals))
@@ -353,19 +354,18 @@ export async function validateTransfer(
     }
 
     // Check that sender and all receivers are NOT on the deny list (if present), or that they are included in the allow list (if present).
-    const senderPromise = token.grpc.getAccountInfo(sender);
-    const receiverPromises = payloads.map((p) => token.grpc.getAccountInfo(p.recipient));
-    const accounts = await Promise.all([senderPromise, ...receiverPromises]);
+    const receiverInfos = await Promise.all(payloads.map((p) => token.grpc.getAccountInfo(p.recipient.address)));
+    const accounts = [senderInfo, ...receiverInfos];
     accounts.forEach((r) => {
         const accToken = r.accountTokens.find((t) => t.id.value === token.info.id.value)?.state;
-        if (accToken?.moduleState === undefined) {
-            return;
-        }
+        if (accToken?.moduleState === undefined)
+            throw new NotAllowedError(TokenHolder.fromAccountAddress(r.accountAddress));
 
-        const moduleState = Cbor.decode(accToken.moduleState) as TokenModuleAccountState;
-        if (moduleState.memberDenyList || moduleState.memberAllowList === false) {
-            throw new NotAllowedError(r.accountAddress);
-        }
+        const accountModuleState = Cbor.decode(accToken.moduleState) as TokenModuleAccountState;
+        if (moduleState.allowList && !accountModuleState.allowList)
+            throw new NotAllowedError(TokenHolder.fromAccountAddress(r.accountAddress));
+        if (moduleState.denyList && accountModuleState.denyList)
+            throw new NotAllowedError(TokenHolder.fromAccountAddress(r.accountAddress));
     });
 
     return true;
@@ -407,8 +407,7 @@ export async function transfer(
     }
 
     if (opts.validate) {
-        // TODO: re-enable validation when it's covered by unit tests
-        // await validateTransfer(token, sender, transfers);
+        await validateTransfer(token, sender, transfers);
     }
 
     const ops: TokenTransferOperation[] = transfers.map((p) => ({ [TokenOperationType.Transfer]: p }));
@@ -427,7 +426,7 @@ export async function transfer(
  */
 export function validateGovernanceOperation(token: Token, sender: AccountAddress.Type): true {
     const { governanceAccount } = Cbor.decode(token.info.state.moduleState) as TokenModuleState;
-    if (!AccountAddress.equals(sender, governanceAccount)) {
+    if (!AccountAddress.equals(sender, governanceAccount.address)) {
         throw new UnauthorizedGovernanceOperationError(sender);
     }
 
@@ -470,8 +469,7 @@ export async function mint(
 
     if (opts.validate) {
         validateGovernanceOperation(token, sender);
-        // TODO: re-enable validation when it's covered by unit tests
-        // amountsList.forEach((amount) => validateAmount(token, amount));
+        amountsList.forEach((amount) => validateAmount(token, amount));
     }
 
     const ops: TokenMintOperation[] = amountsList.map((amount) => ({
@@ -509,8 +507,7 @@ export async function burn(
 
     if (opts.validate) {
         validateGovernanceOperation(token, sender);
-        // TODO: re-enable validation when it's covered by unit tests
-        // amountsList.forEach((amount) => validateAmount(token, amount));
+        amountsList.forEach((amount) => validateAmount(token, amount));
     }
 
     const ops: TokenBurnOperation[] = amountsList.map((amount) => ({
@@ -529,7 +526,7 @@ type UpdateListOptions = {
  *
  * @param {Token} token - The token for which to add the list entry.
  * @param {AccountAddress.Type} sender - The account address of the sender.
- * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
+ * @param {TokenHolder.Type | TokenHolder.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
  * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
@@ -540,7 +537,7 @@ type UpdateListOptions = {
 export async function addAllowList(
     token: Token,
     sender: AccountAddress.Type,
-    targets: AccountAddress.Type | AccountAddress.Type[],
+    targets: TokenHolder.Type | TokenHolder.Type[],
     signer: AccountSigner,
     expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
     { validate }: UpdateListOptions = { validate: true }
@@ -560,7 +557,7 @@ export async function addAllowList(
  *
  * @param {Token} token - The token for which to add the list entry.
  * @param {AccountAddress.Type} sender - The account address of the sender.
- * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
+ * @param {TokenHolder.Type | TokenHolder.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
  * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
@@ -571,7 +568,7 @@ export async function addAllowList(
 export async function removeAllowList(
     token: Token,
     sender: AccountAddress.Type,
-    targets: AccountAddress.Type | AccountAddress.Type[],
+    targets: TokenHolder.Type | TokenHolder.Type[],
     signer: AccountSigner,
     expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
     { validate }: UpdateListOptions = { validate: true }
@@ -591,7 +588,7 @@ export async function removeAllowList(
  *
  * @param {Token} token - The token for which to add the list entry.
  * @param {AccountAddress.Type} sender - The account address of the sender.
- * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
+ * @param {TokenHolder.Type | TokenHolder.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
  * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
@@ -602,7 +599,7 @@ export async function removeAllowList(
 export async function addDenyList(
     token: Token,
     sender: AccountAddress.Type,
-    targets: AccountAddress.Type | AccountAddress.Type[],
+    targets: TokenHolder.Type | TokenHolder.Type[],
     signer: AccountSigner,
     expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
     { validate }: UpdateListOptions = { validate: true }
@@ -622,7 +619,7 @@ export async function addDenyList(
  *
  * @param {Token} token - The token for which to add the list entry.
  * @param {AccountAddress.Type} sender - The account address of the sender.
- * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
+ * @param {TokenHolder.Type | TokenHolder.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
  * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
@@ -633,7 +630,7 @@ export async function addDenyList(
 export async function removeDenyList(
     token: Token,
     sender: AccountAddress.Type,
-    targets: AccountAddress.Type | AccountAddress.Type[],
+    targets: TokenHolder.Type | TokenHolder.Type[],
     signer: AccountSigner,
     expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
     { validate }: UpdateListOptions = { validate: true }