+++ packages/sdk/src/pub/plt.ts
@@ -1,7 +1,6 @@
 export * from '../plt/types.js';
 export * from '../plt/module.js';
 
-export * as TokenMetadataUrl from '../plt/TokenMetadataUrl.js';
 export * as Token from '../plt/Token.js';
 
 // To limit the exports meant only for internal use, we re-create the module exports.
@@ -10,3 +9,5 @@ export * as Cbor from './plt/Cbor.js';
 export * as CborMemo from './plt/CborMemo.js';
 export * as TokenId from './plt/TokenId.js';
 export * as TokenModuleReference from './plt/TokenModuleReference.js';
+export * as TokenMetadataUrl from './plt/TokenMetadataUrl.js';
+export * as TokenHolder from './plt/TokenHolder.js';